% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_tools.R
\name{addCurveweights}
\alias{addCurveweights}
\title{Add slingshot curve weights to Seurat object}
\usage{
addCurveweights(seuratObj, slingshotObj)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{slingshotObj}{A PseudotimeOrdering object generated using
\code{slingshot::slingshot}}
}
\value{
A Seurat object with slingshot curve weights added to the metadata.
}
\description{
This function adds slingshot lineages to Seurat object.
}
