% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_object.R
\name{addMetadataCategory}
\alias{addMetadataCategory}
\title{Add a categorical column to a Seurat metadata or SingleCellExperiment
coldata}
\usage{
addMetadataCategory(
  scObj,
  col,
  newCol,
  keys,
  values,
  newCol2 = NULL,
  values2 = NULL
)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{col}{Column whose values will be used for creating the new column.}

\item{newCol}{Column to be added.}

\item{keys}{A list of keys. If vectors are part of the keys,
each of their elements will be assigned the corresponding value.}

\item{values}{A vector of values.
Must have the same length as \code{keys}.}

\item{newCol2}{A second column to be added based on the same keys.
Default is \code{NULL} (no second column will be added).}

\item{values2}{A vector of values corresponding to the second
column. Default is \code{NULL} (no second column will be added).}
}
\value{
A \code{Seurat} or \code{SingleCellExpression} object with one or
two new categorical column(s) in the metadata/coldata.
}
\description{
Add a categorical column to a Seurat metadata or SingleCellExperiment
coldata
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs2', package='hammers')
sceObj <- qs2::qs_read(scePath)
sceObj <- addMetadataCategory(sceObj, 'Cell_Cycle', 'Type',
list(c('G0', 'G1'), 'G2M', 'S'), c(2, 3, 1))

}
