% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_tools.R
\name{addSlingshotResults}
\alias{addSlingshotResults}
\title{Add slingshot information to Seurat object.}
\usage{
addSlingshotResults(seuratObj, slingshotObj, colStr, fun)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{slingshotObj}{A PseudotimeOrdering object generated using
\code{slingshot::slingshot}}

\item{colStr}{Column name}

\item{fun}{Function}
}
\value{
A Seurat object with slingshot information added to the metadata.
}
\description{
This function adds slingshot information to Seurat object.
}
\keyword{internal}
