% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{bfCorrectDF}
\alias{bfCorrectDF}
\title{Perform multiple testing correction and filtering with Bonferroni}
\usage{
bfCorrectDF(df, nTests, pvalThr = 0.05, colStr = "pval")
}
\arguments{
\item{df}{A dataframe with a column of p-values.}

\item{nTests}{Number of tests.}

\item{pvalThr}{p-value threshold.}

\item{colStr}{Name of the column of p-values.}
}
\value{
The data frame with Benjamini-Yekutieli-corrected p-values.
}
\description{
This function performs the Bonferroni correction for multiple
testing in a dataframe column of p-values and filters the data-frame
based on p-values.
}
\examples{
df <- data.frame(elem = c('A', 'B', 'C', 'D', 'E'),
pval = c(0.032, 0.001, 0.0045, 0.051, 0.048))
bfCorrectDF(df, 5)

}
