% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{bfCorrectV}
\alias{bfCorrectV}
\title{Perform the Bonferroni multiple testing correction on a vector of p-values}
\usage{
bfCorrectV(pvals, nTests, replaceHigh = TRUE)
}
\arguments{
\item{pvals}{A vector of p-values.}

\item{nTests}{Number of tests.}

\item{replaceHigh}{Whether to replace adjusted p-values greater than 1
with 1.}
}
\value{
A vector of Bonferroni-corrected p-values.
}
\description{
This function performs the Bonferroni multiple testing correction on a
vector of p-values.
}
\keyword{internal}
