% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{bhCorrectDF}
\alias{bhCorrectDF}
\title{Perform multiple testing correction and filtering with Benjamini-Hochberg}
\usage{
bhCorrectDF(df, ...)
}
\arguments{
\item{df}{A dataframe with a column of p-values.}

\item{...}{Additional arguments passed to \code{fdrCorrectDF}.}
}
\value{
The data frame with Benjamini-Hochberg-corrected p-values.
}
\description{
This function performs the Benjamini-Hochberg correction for multiple
testing in a dataframe column of p-values and filters the data-frame
based on p-values.
}
\examples{
df <- data.frame(elem = c('A', 'B', 'C', 'D', 'E'),
pval = c(0.032, 0.001, 0.0045, 0.051, 0.048))
bhCorrectDF(df)

}
