% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{bhCorrectV}
\alias{bhCorrectV}
\title{Perform the Benjamini-Hochberg multiple testing correction on a vector of
p-values}
\usage{
bhCorrectV(pvals, keepInputOrder = TRUE)
}
\arguments{
\item{pvals}{A vector of p-values.}

\item{keepInputOrder}{Whether to keep the input order of p-values as opposed
to returning them as sorted.}
}
\value{
A vector of Benjamini-Hochberg-corrected p-values.
}
\description{
This function performs the Benjamini-Hochberg multiple testing correction on
a vector of p-values.
}
\keyword{internal}
