% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_of_mass.R
\name{centerOfMass}
\alias{centerOfMass}
\title{Calculate the center of mass of columns}
\usage{
centerOfMass(obj, weights)
}
\arguments{
\item{obj}{A data frame or matrix.}

\item{weights}{A vector of weights.}
}
\value{
A vector containing the center of mass.
}
\description{
This function calculates the center of mass based on the columns of
a data frame or matrix and a vector of weights.
}
\examples{
obj <- matrix(data=c(2, 3, 1, 3, 6, 8), nrow=3, ncol=2)
weights <- c(0.8, 6, 16)
centerOfMass(obj, weights)

}
