% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkGenes}
\alias{checkGenes}
\title{Check if all genes exist in the single-cell expression object}
\usage{
checkGenes(scObj, genes)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{genes}{A character vector of genes.}
}
\value{
None. This function is called for its side effect.
}
\description{
This function checks if all genes exist in the single-cell expression
object.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
checkGenes(scObj, c('Gene_0980', 'Gene_0981', 'Gene_0982'))

}
