% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_of_mass.R
\name{colCenters}
\alias{colCenters}
\title{Calculate the centers of mass of metadata/coldata columns}
\usage{
colCenters(scObj, columns)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{columns}{Numeric columns.}
}
\value{
A data frame containing the coordinates of centers of mass.
}
\description{
This function calculates the centers of mass of  selected metadata/coldata
columns from a \code{Seurat} or \code{SingleCellExpression}
object.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scObj <- scuttle::logNormCounts(scObj)
scObj <- scater::runUMAP(scObj)
colCenters(scObj, c('sizeFactor'))

}
