% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{colsDimPlot}
\alias{colsDimPlot}
\title{Create a single-cell dimensionality reduction plot with added labeled points
for numeric columns}
\usage{
colsDimPlot(scObj, cols, ...)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{cols}{Columns whose centers of mass will be plotted.}

\item{...}{Additional parameters passed to \code{pointsDimPlot}.}
}
\value{
A ggplot object.
}
\description{
This function creates a single-cell dimensionality reduction plot with
added labeled points for metadata numeric columns.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs2', package='hammers')
sceObj <- qs2::qs_read(scePath)
colsDimPlot(sceObj, c('sizeFactor', 'silhouette'))

}
