% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{countsVector}
\alias{countsVector}
\title{Extract count information from Seurat column.}
\usage{
countsVector(seuratObj, column)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{column}{Column}
}
\value{
A frequency vector with the unique column values as names.
}
\description{
This function extracts count information from Seurat column.
}
