% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/visualization.R
\name{devPlot}
\alias{devPlot}
\alias{devPlot.default}
\alias{devPlot.function}
\alias{devPlot.ggplot}
\alias{devPlot.list}
\title{Saves plot or list of plots}
\usage{
devPlot(plotObject, ...)

\method{devPlot}{default}(plotObject, ...)

\method{devPlot}{`function`}(plotObject, ...)

\method{devPlot}{ggplot}(plotObject, ...)

\method{devPlot}{list}(plotObject, ...)
}
\arguments{
\item{plotObject}{A function, ggplot object, or list of ggplot objects.}

\item{...}{Additional arguments.}
}
\value{
No value. This function is called for its side effect.
}
\description{
This function saves a plot or list of plots as a pdf. Can also take as input
a function that returns a ggplot object together with its arguments.
}
\examples{
library(ggplot2)
df <- data.frame(x = c(1, 2), y = c(3, 5))
p <- ggplot(df) + geom_point(aes(x, y))
devPlot(p)

simplePlot <- function(df, title)
    return(ggplot(df) + geom_point(aes(x, y)) + ggtitle(title))

devPlot(simplePlot, df, 'Plot title')

}
