% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{distributionPlot}
\alias{distributionPlot}
\title{Plot the distribution of cells across two columns}
\usage{
distributionPlot(
  scObj,
  title = NULL,
  col1 = "seurat_clusters",
  col2 = "orig.ident",
  type = c("counts", "percs"),
  xLab = col1,
  yLab = if (type == "counts") "Count" else "Percentage",
  legendLab = col2,
  palette = "Spectral",
  legendPos = "right",
  legendTextSize = 10,
  legendTitleSize = 10,
  axisTextSize = 12,
  axisTitleSize = 12,
  sigDigits = 2
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{title}{Plot title.}

\item{col1}{Column as string.}

\item{col2}{Column as string.}

\item{type}{Whether the plot should display counts ('counts', default) or
percentages ('percs').}

\item{xLab}{x axis label.}

\item{yLab}{y axis label.}

\item{legendLab}{Legend label.}

\item{palette}{Color palette.}

\item{legendPos}{Legend position.}

\item{legendTextSize}{Legend text size.}

\item{legendTitleSize}{Legend title size.}

\item{axisTextSize}{Axis text size.}

\item{axisTitleSize}{Axis title size.}

\item{sigDigits}{Number of significant digits used by percentages displayed
on the plot. Ignored if \code{type} is 'counts'.}
}
\value{
A ggplot object.
}
\description{
This function plots the distribution of cells across two columns.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
p <- distributionPlot(sceObj, col1='Cluster', col2='Donor')

}
