% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{entrezGenes}
\alias{entrezGenes}
\title{Convert gene symbols to Entrez IDs.}
\usage{
entrezGenes(genes, species = c("human", "mouse", "zebrafish"))
}
\arguments{
\item{genes}{A character vector of gene symbols.}

\item{species}{Species. Must be one of 'human', 'mouse' and 'zebrafish'.}
}
\value{
The Entrez IDs of the genes.
}
\description{
This function converts gene symbols to Entrez IDs.
}
\keyword{internal}
