% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{fdrCorrectDF}
\alias{fdrCorrectDF}
\title{Perform multiple testing correction by controlling the false discovery rate}
\usage{
fdrCorrectDF(
  df,
  fdrControlFun,
  pvalThr = 0.05,
  colStr = "pval",
  newColStr = "pvalAdj"
)
}
\arguments{
\item{df}{A dataframe with a column of p-values.}

\item{pvalThr}{p-value threshold.}

\item{colStr}{Name of the column of p-values.}

\item{newColStr}{Name of the column of adjusted p-values that will be
created.}
}
\value{
The data frame with p-values corrected using the method of choice
(Benjamini-Hochberg or Benjamini-Yekutieli)
}
\description{
This function perform multiple testing correction by controlling the false
discovery rate.
}
\keyword{internal}
