% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_genes.R
\name{findRareGenes}
\alias{findRareGenes}
\title{Find rare genes in a Seurat or SingleCellExpression object}
\usage{
findRareGenes(scObj, nCells = 10)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{nCells}{Minimum number of cells in which a gene must be expressed to
be regarded as non-rare.}
}
\value{
A data frame with the rare genes as rownames and a single column
representing their frequencies.
}
\description{
This function finds genes expressed in a low number of cells in a Seurat or
SingleCellExpression object.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
df <- findRareGenes(sceObj)

}
