% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_information.R
\name{geneCellSets}
\alias{geneCellSets}
\title{Generates cell sets for each input gene}
\usage{
geneCellSets(scObj, genes = NULL)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{genes}{Selected genes. If \code{NULL}, all genes will be retained.}
}
\value{
A named list of character vectors of cell names.
}
\description{
This function constructs, for each input gene, sets of cells expressing
the gene
}
\examples{
mat <- matrix(0, 1000, 500)
rownames(mat) <- paste0('G', seq(1000))
colnames(mat) <- paste0('C', seq(500))
mat[sample(length(mat), 70000)] <- sample(50, 70000, TRUE)
mat <- mat[paste0('G', sample(1000, 3)), ]
geneCellSets(mat)

}
