% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_information.R
\name{geneCounts}
\alias{geneCounts}
\title{Extract gene counts from a Seurat or SingleCellExperiment object}
\usage{
geneCounts(scObj, minCutoff = NULL, maxCutoff = NULL)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{minCutoff}{Maximum cutoff for gene counts. Genes with counts above
this value will be omitted.}
}
\value{
A data frame of gene counts
}
\description{
This function extracts gene counts from a Seurat or
SingleCellExperiment object as a data frame.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=200))
df <- geneCounts(scObj)

}
