% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_information.R
\name{genePresence}
\alias{genePresence}
\title{Extract gene presence from a Seurat or SingleCellExperiment object}
\usage{
genePresence(scObj, genes = NULL, minCutoff = NULL, maxCutoff = NULL)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{genes}{Genes for which the number of cells in which the gene is
expressed will be computed. If \code{NULL} (as default), all genes will be
included in the assessment.}

\item{minCutoff}{Minimum cutoff for gene counts. Genes with counts below
this value will be omitted.}

\item{maxCutoff}{Maximum cutoff for gene counts. Genes with counts above
this value will be omitted.}
}
\value{
A data frame with two columns. The first column lists the genes
ordered decreasingly by the number of cells in which they appear, the second
lists the corresponding numbers of cells.
}
\description{
This function extracts the number of cells in which a gene from a Seurat or
SingleCellExperiment is expressed.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
df <- genePresence(sceObj)

}
