% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{genesDimPlot}
\alias{genesDimPlot}
\title{Create a single-cell dimensionality reduction plot with added labeled points
for genes}
\usage{
genesDimPlot(scObj, genes, ...)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{genes}{Genes whose centers of mass will be plotted.}

\item{...}{Additional parameters passed to \code{pointsDimPlot}.}
}
\value{
A ggplot object.
}
\description{
This function creates a single-cell dimensionality reduction plot with
added labeled points for genes.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs2', package='hammers')
sceObj <- qs2::qs_read(scePath)
genesDimPlot(sceObj, c('Gene_0364', 'Gene_0388', 'Gene_0477'))

}
