% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{genesER}
\alias{genesER}
\title{Perform enrichment analysis on a set of genes}
\usage{
genesER(
  genes,
  species,
  funString = c("enrichGO", "enrichKEGG", "enrichWP"),
  pvalThr = 0.05,
  addNegLog = FALSE
)
}
\arguments{
\item{genes}{A character vector of gene symbols.}

\item{species}{Species. Must be one of 'human', 'mouse' and 'zebrafish'.}

\item{funString}{Name of enrichment function from \code{clusterProfiler}.
Must be a character selected from 'enrichGO', 'enrichKEGG' and 'enrichWP'.}

\item{pvalThr}{p-value threshold.}

\item{addNegLog}{Whether to compute the negative log10 of the adjusted
p-value.}
}
\value{
Enrichment result.
}
\description{
This function performs enrichment analysis on a set of genes.
}
\examples{
m <- genesER(c('AURKA', 'TOP2A', 'CENPF', 'PTTG2', 'MKI67', 'BIRC5', 'RRM2'),
'human')

}
