% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{getEnrichmentResult}
\alias{getEnrichmentResult}
\title{Perform enrichment analysis on a set of Entrez IDs}
\usage{
getEnrichmentResult(
  entrezIDs,
  species = c("human", "mouse", "zebrafish"),
  funString = c("enrichGO", "enrichKEGG", "enrichWP")
)
}
\arguments{
\item{entrezIDs}{A character vector of gene Entrez IDs.}

\item{species}{Species. Must be one of 'human', 'mouse' and 'zebrafish'.}

\item{funString}{Name of enrichment function from \code{clusterProfiler}.
Must be a character selected from 'enrichGO', 'enrichKEGG' and 'enrichWP'.}
}
\value{
Enrichment result.
}
\description{
This function performs enrichment analysis on a set of Entrez IDs.
}
\keyword{internal}
