% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammers-package.R
\docType{package}
\name{hammers-package}
\alias{hammers}
\alias{hammers-package}
\title{hammers: Utilities for scRNA-seq data analysis}
\description{
hammers is a utilities suite for scRNA-seq data analysis compatible with both Seurat and SingleCellExperiment. It provides simple tools to address tasks such as retrieving aggregate gene statistics, finding and removing rare genes, performing representation analysis, computing the center of mass for the expression of a gene of interest in low-dimensional space, and calculating silhouette and cluster-normalized silhouette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrei-stoica26/hammers}
  \item Report bugs at \url{https://github.com/andrei-stoica26/hammers/issues}
}

}
\author{
\strong{Maintainer}: Andrei-Florian Stoica \email{andreistoica@foxmail.com} (\href{https://orcid.org/0000-0002-5253-0826}{ORCID})

}
\keyword{internal}
