% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_object.R
\name{hashMap}
\alias{hashMap}
\title{Create a hash map}
\usage{
hashMap(hashKeys, hashValues)
}
\arguments{
\item{hashKeys}{A list of hash keys. If vectors are part of the hash keys,
each of their elements will be assigned the corresponding value.}

\item{hashValues}{A vector of hash values.
Must have the same length as \code{hashKeys}.}
}
\value{
A named vector.
}
\description{
This function creates a hash map.
}
\examples{
hashMap(list(2, c(3, 4, 5), 6, 8), c('a', 'b', 'c', 'd'))

}
