% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{joinCharCombs}
\alias{joinCharCombs}
\title{Join all combinations of elements from character vectors}
\usage{
joinCharCombs(..., joinChar = "_")
}
\arguments{
\item{...}{Vectors passed to \code{expandGrid}.}

\item{joinChar}{Character used to join combinations.}
}
\value{
A character vector.
}
\description{
This function joins all combinations of elements from character vectors
with a separating character.
}
