% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{joinER}
\alias{joinER}
\title{Join the data frames of two enrichment results}
\usage{
joinER(
  er1,
  er2,
  sharedCols = c("Description", "BgRatio"),
  specificCols = c("GeneRatio", "p.adjust", "geneID", "Count")
)
}
\arguments{
\item{er1}{An \code{enrichResult} object.}

\item{er2}{An \code{enrichResult} object.}

\item{sharedCols}{Columns from the \code{enrichResult} objects used for the
join that have the same values for each term shared by the two objects.}

\item{specificCols}{Columns from the \code{enrichResult} objects used for the
join that have potentially different values for terms shared by the
two objects.}
}
\value{
Genes enriched for terms.
}
\description{
This function joins the data frames of two \code{enrichResult} objects.
}
\examples{
m1 <- genesER(c('AURKA', 'PTTG2', 'MKI67', 'RRM2'),
'human')
m2 <- genesER(c('AURKA', 'TOP2A', 'CENPF', 'BIRC5'),
'human')
df <- joinER(m1, m2)


}
