% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_object.R
\name{keyvalMap}
\alias{keyvalMap}
\title{Create a map from keys to values}
\usage{
keyvalMap(keys, values)
}
\arguments{
\item{keys}{A list of keys. If vectors are part of the keys,
each of their elements will be assigned the corresponding value.}

\item{values}{A vector of values.
Must have the same length as \code{keys}.}
}
\value{
A named vector.
}
\description{
This function creates a map from keys to values.
}
\examples{
keyvalMap(list(2, c(3, 4, 5), 6, 8), c('a', 'b', 'c', 'd'))

}
