% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R, R/generics.R
\name{metadataDF.default}
\alias{metadataDF.default}
\alias{metadataDF.Seurat}
\alias{metadataDF.SingleCellExperiment}
\alias{metadataDF}
\title{Extract metadata from object as a data frame}
\usage{
\method{metadataDF}{default}(scObj)

\method{metadataDF}{Seurat}(scObj)

\method{metadataDF}{SingleCellExperiment}(scObj)

metadataDF(scObj)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}
}
\value{
A metadata data frame.
}
\description{
This function extracts the metadata from a Seurat or
SingleCellExperiment object as a data frame
}
