% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R, R/generics.R
\name{metadataNames.default}
\alias{metadataNames.default}
\alias{metadataNames.Seurat}
\alias{metadataNames.SingleCellExperiment}
\alias{metadataNames}
\title{Return metadata names.}
\usage{
\method{metadataNames}{default}(scObj)

\method{metadataNames}{Seurat}(scObj)

\method{metadataNames}{SingleCellExperiment}(scObj)

metadataNames(scObj)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}
}
\value{
The names of the metadata columns.
}
\description{
This function extracts metadata names from a Seurat or
SingleCellExperiment object.
}
