% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{mtCorrectDF}
\alias{mtCorrectDF}
\title{Perform multiple testing correction on a data frame column}
\usage{
mtCorrectDF(
  df,
  mtMethod = c("bf", "bh", "by"),
  pvalThr = 0.05,
  colStr = "pval",
  newColStr = "pvalAdj",
  ...
)
}
\arguments{
\item{df}{A data frame with a p-values columnn.}

\item{mtMethod}{Multiple testing correction method. Choices are
Bonferroni ('bf'), Benjamini-Hochberg('bh'), and Benjamini-Yekutieli ('by').}

\item{pvalThr}{p-value threshold.}

\item{colStr}{Name of the column of p-values.}

\item{newColStr}{Name of the column of adjusted p-values that will be
created.}

\item{...}{Additional arguments passed to the multiple testing correction
method.}
}
\value{
A data frame with the p-value column corrected for multiple testing.
}
\description{
This function orders a data frame based on a column of p-values, performs
multiple testing on the column, and filters the data-frame based on it.
}
\examples{
df <- data.frame(elem = c('A', 'B', 'C', 'D', 'E'),
pval = c(0.032, 0.001, 0.0045, 0.051, 0.048))
mtCorrectDF(df)

}
