% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{mtCorrectV}
\alias{mtCorrectV}
\title{Perform multiple testing correction with Benjamini-Yekutieli on a vector of
p-values}
\usage{
mtCorrectV(
  pvals,
  mtMethod = c("bf", "bh", "by"),
  mtStat = c("identity", "median", "mean", "max", "min"),
  ...
)
}
\arguments{
\item{pvals}{A vector of p-values.}

\item{mtMethod}{Multiple testing correction method. Choices are
Bonferroni ('bf'), Benjamini-Hochberg('bh'), and Benjamini-Yekutieli ('by').}

\item{mtStat}{A statistics to be optionally computed. Choices are 'identity'
(no statistics will be computed and the adjusted p-values will be returned
as such), 'median', 'mean', 'max' and 'min'.}

\item{...}{Additional parameters passed to the multiple testing function.}
}
\value{
A vector of p-values corrected for multiple testing.
}
\description{
This function perform multiple testing correction with Benjamini-Yekutieli
on a vector of p-values.
}
\examples{
pvals <- c(0.032, 0.001, 0.0045, 0.051, 0.048)
mtCorrectV(pvals)

}
