% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearestNeighbors}
\alias{nearestNeighbors}
\title{Get nearest neighbors from distance matrix}
\usage{
nearestNeighbors(distMat)
}
\arguments{
\item{distMat}{A distance matrix.}
}
\value{
A named character vector.
}
\description{
This function gets the nearest neighbors from a distance matrix.
}
\examples{
df <- data.frame(v = c(1, 2, 4, 5, 6),
w = c(2, 3, 1, 5, 8),
x = c(2, 8, 7, 1, 1),
y = c(2, 3, 2, 2, 4),
z = c(1, 9, 9, 7, 6))
distMat <- as.matrix(stats::dist(df))
rownames(distMat) <- c('v', 'w', 'x', 'y', 'z')
colnames(distMat) <- c('v', 'w', 'x', 'y', 'z')
nearestNeighbors(distMat)

}
