% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{normalizeSilhouette}
\alias{normalizeSilhouette}
\title{Normalize silhouette by identity class for single-cell expression object}
\usage{
normalizeSilhouette(scObj, idClass, silCol = "silhouette")
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{idClass}{Identity class. Must be present among the metadata columns of
the single-cell expression object.}

\item{silCol}{The name of the silhouette column.}
}
\value{
A data frame with normalized silhouettes for each unique element in
the identity class.
}
\description{
This function normalizes the already computed silhouette for each identity
class in the single-cell expression object.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
df <- normalizeSilhouette(sceObj, 'Cluster')

}
