% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numCosine}
\alias{numCosine}
\title{Calculate the cosine similarity of two vectors}
\usage{
numCosine(x, y = NULL)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{y}{A numeric vector of the same dimension as \code{x} if the latter
is a vector. If \code{NULL}, the similarity matrix between all column
vectors of \code{x} will be computed.}
}
\value{
A numeric values representing the cosine similarity of the two
vectors.
}
\description{
This function calculates the cosine similarity of two vectors.
}
\examples{
numCosine(c(2, 3, 6), c(4, 3, 2))

}
