% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pointsDimPlot}
\alias{pointsDimPlot}
\title{Create a single-cell dimensionality reduction plot with added labeled points}
\usage{
pointsDimPlot(
  scObj,
  title = NULL,
  pointsObj = NULL,
  alpha = 1,
  pointShape = 4,
  pointSize = 2,
  pointColor = "black",
  labelSize = 2.5,
  maxOverlaps = 30,
  ...
)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{title}{Plot title.}

\item{pointsObj}{A data frame or matrix of points with two columns
representing x and y coordinates.}

\item{alpha}{Opaqueness level.}

\item{pointShape}{Point shape.}

\item{pointSize}{Point size.}

\item{pointColor}{Point color.}

\item{labelSize}{Label size. If \code{NULL}, the points will not receive
labels.}

\item{maxOverlaps}{Maximum overlaps.}

\item{...}{Additional parameters passed to \code{Seurat::DimPlot}.}
}
\value{
A ggplot object.
}
\description{
This function creates a single-cell dimensionality reduction plot with
added labeled points.
}
\details{
A wrapper around \code{scLang::dimPlot}.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs2', package='hammers')
sceObj <- qs2::qs_read(scePath)
pointsObj <- data.frame(x = c(2, 3),
y = c(1, 0),
row.names = c('P1', 'P2'))
pointsDimPlot(sceObj, pointsObj=pointsObj)

}
