% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_aux.R
\name{prepAlluvial}
\alias{prepAlluvial}
\title{Prepare dataframe for alluvial plot}
\usage{
prepAlluvial(
  df,
  pvalCol = "pvalAdj",
  colIndices = c(1, 2),
  weightExp = 1/2,
  pvalOffset = 1.00000023069254e-317
)
}
\arguments{
\item{df}{A data frame.}

\item{pvalCol}{Name of p-value column to be used by the alluvial plot.}

\item{colIndices}{A vector respresenting the indices of the two categorical
columns from the data frame that will be used.}

\item{weightExp}{Exponent used in constructing weight from p-values.}

\item{pvalOffset}{Offset used to avoid zeros inside the logarithm function.}
}
\value{
A data frame with weight scores in lieu of p-values.
}
\description{
This function extracts the relevant information from a data frame and
adjusts p-values to be used as weights for the alluvia.
}
\examples{
df <- data.frame(A = c('a1', 'a2', 'a3', 'a4'),
B = c('b1', 'b2', 'b3', 'b4'),
pvalAdj = c(0.81, 1e-6, 1e-3, 0.022))
prepAlluvial(df)

}
