% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proximity}
\alias{proximity}
\title{Compute proximity between two vectors based on Euclidean distance}
\usage{
proximity(x, y, maxDist)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{maxDist}{Maximum distance.}
}
\value{
A number between 0 and 1.
}
\description{
This functions computes proximity between two vectors based on Euclidean
distance and an input maximum distance.
}
\examples{
proximity(2, 3, 6)

}
