% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pvalRiverPlot}
\alias{pvalRiverPlot}
\title{Plot representation data frame}
\usage{
pvalRiverPlot(df, weightExp = 1/2, ...)
}
\arguments{
\item{df}{A data frame.}

\item{weightExp}{Exponent used in constructing weight from p-values.}

\item{...}{Additional parameters passed to \code{henna::riverPlot}}
}
\value{
A ggplot object
}
\description{
This function plots representation data frame as an alluvial plot.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scCol(scObj, 'Cluster') <- withr::with_seed(1,
sample(paste0('Cluster', seq(5)), dim(scObj)[2], replace=TRUE))
scCol(scObj, 'Donor') <- rep('Donor1', dim(scObj)[2])
for (i in seq(5)){
scCol(scObj, 'Donor')[sample(which(scCol(scObj, 'Cluster') ==
paste0('Cluster', i)), 15)]<- paste0('Donor', sample(seq(6), 1))
scCol(scObj, 'Donor')[sample(which(scCol(scObj, 'Cluster') ==
paste0('Cluster', i))
, 15)]<- paste0('Donor', sample(seq(6), 1))
}
df <- repAnalysis(scObj, 'Cluster', 'Donor')
pvalRiverPlot(df)

}
