% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{repAnalysis}
\alias{repAnalysis}
\title{Find the differential representation of two Seurat or SingleCellExperiment
columns}
\usage{
repAnalysis(
  scObj,
  col1 = "seurat_clusters",
  col2 = "orig.ident",
  doOverrep = TRUE,
  mtMethod = c("BY", "holm", "hochberg", "hommel", "bonferroni", "BH", "fdr", "none"),
  ...
)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{col1}{Column as string.}

\item{col2}{Column as string.}

\item{doOverrep}{Whether to perform overrepresentation analysis. If
\code{FALSE}, underrepresentation analysis will be performed instead.}

\item{mtMethod}{Multiple testing correction method. Choices are
'BY' (default) 'holm', hochberg', hommel', 'bonferroni', 'BH',  'fdr' and
'none'.}

\item{...}{Additional parameters passed to \code{LISTO::mtCorrectDF}.}
}
\value{
An overrepresentation or underrepresentation data frame.
}
\description{
This function finds the differential representation of two
\code{Seurat} or \code{SingleCellExperiment} columns.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs2', package='hammers')
sceObj <- qs2::qs_read(scePath)
repAnalysis(sceObj, 'Cluster', 'Donor')

}
