% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{repPlot}
\alias{repPlot}
\title{Plot representation data frame}
\usage{
repPlot(df, ...)
}
\arguments{
\item{df}{A representation data frame.}

\item{...}{Additional parameters passed to \code{henna::riverPlot}}
}
\value{
A ggplot object
}
\description{
This function plots representation data frame as an alluvial plot.
}
