% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeMessage}
\alias{safeMessage}
\title{Message an input if verbose is set to TRUE}
\usage{
safeMessage(msg, verbose)
}
\arguments{
\item{msg}{Message}

\item{verbose}{Whether the message should be displayed.}
}
\value{
No return value. This function is called for its side effect
(messaging the input if \code{verbose} is set to \code{TRUE}).
}
\description{
This function messages an input if \code{verbose} is set to \code{TRUE}.
}
