% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeMinmax}
\alias{safeMinmax}
\title{Perform min-max normalization when possible; otherwise return a zero-vector}
\usage{
safeMinmax(scores, safeVal = 0)
}
\arguments{
\item{scores}{Numeric vector}

\item{safeVal}{Value to replace all values with when all values in the vector
are the same}
}
\value{
Min-max-normalized scores or the zero vector
}
\description{
This function min-max-normalizes a vector when possible, and otherwise returns
the zero vector
}
