% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R, R/generics.R
\name{scCol.default}
\alias{scCol.default}
\alias{scCol.Seurat}
\alias{scCol.SingleCellExperiment}
\alias{scCol}
\title{Extract a metadata column from object.}
\usage{
\method{scCol}{default}(scObj, colStr)

\method{scCol}{Seurat}(scObj, colStr)

\method{scCol}{SingleCellExperiment}(scObj, colStr)

scCol(scObj, colStr)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{colStr}{Column name.}
}
\value{
A column vector.
}
\description{
This function extracts a metadata column from a Seurat or
SingleCellExperiment object.
}
