% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{scColPairCounts}
\alias{scColPairCounts}
\title{Extract per-pair counts from single-cell expression object columns}
\usage{
scColPairCounts(scObj, col1 = "seurat_clusters", col2 = "orig.ident")
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{col1}{Column as string.}

\item{col2}{Column as string.}
}
\value{
A data frame listing the counts of all combinations of pairs from
two categorical columns.
}
\description{
This function extracts per-pair counts from two columns of a single-cell
expression object.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
scColPairCounts(sceObj, 'Mutation_Status', 'Cell_Cycle')

}
