% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R, R/generics.R
\name{scExpMat.default}
\alias{scExpMat.default}
\alias{scExpMat.Seurat}
\alias{scExpMat.SingleCellExperiment}
\alias{scExpMat}
\title{Extracts the expression matrix from object.}
\usage{
\method{scExpMat}{default}(
  scObj,
  dataType = c("counts", "data", "logcounts"),
  genes = NULL,
  densify = TRUE
)

\method{scExpMat}{Seurat}(
  scObj,
  dataType = c("counts", "data", "logcounts"),
  genes = NULL,
  densify = TRUE
)

\method{scExpMat}{SingleCellExperiment}(
  scObj,
  dataType = c("counts", "data", "logcounts"),
  genes = NULL,
  densify = TRUE
)

scExpMat(
  scObj,
  dataType = c("counts", "data", "logcounts"),
  genes = NULL,
  densify = TRUE
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{dataType}{Expression data type.}

\item{genes}{Selected genes. If \code{NULL}, all genes will be retained}

\item{densify}{Whether to convert to dense matrix.}
}
\value{
A PCA matrix.
}
\description{
This function extracts an expression matrix from a Seurat or
SingleCellExperiment object.
}
