% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{scPCAMat}
\alias{scPCAMat}
\title{Extracts the PCA matrix from object.}
\usage{
scPCAMat(scObj)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}
}
\value{
A PCA matrix.
}
\description{
This function extracts the PCA matrix from a Seurat or
SingleCellExperiment object.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scObj <- scuttle::logNormCounts(scObj)
scObj <- scater::runPCA(scObj)
pcaMat <- scPCAMat(scObj)

}
