% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{scUMAPMat}
\alias{scUMAPMat}
\title{Extracts the UMAP matrix from object.}
\usage{
scUMAPMat(scObj)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}
}
\value{
A UMAP matrix.
}
\description{
This function extracts the UMAP matrix from a Seurat or
SingleCellExperiment object.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scObj <- scuttle::logNormCounts(scObj)
scObj <- scater::runUMAP(scObj)
umapMat <- scUMAPMat(scObj)

}
