% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{termGenes}
\alias{termGenes}
\title{Extract genes enriched for term}
\usage{
termGenes(er, term)
}
\arguments{
\item{er}{Enrichment result.}

\item{term}{Term.}
}
\value{
Genes enriched for term.
}
\description{
This function extracts genes enriched for term from an \code{enrichResult}
object.
}
