% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{termGenesHelper}
\alias{termGenesHelper}
\title{Extract genes enriched for terms}
\usage{
termGenesHelper(er, terms)
}
\arguments{
\item{er}{Enrichment result.}

\item{terms}{Terms for which enriched genes should be extracted.}
}
\value{
Genes enriched for terms.
}
\description{
This function extracts genes enriched for term from an \code{enrichResult}
object.
}
\keyword{internal}
