% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_and_memory.R
\name{timeExpr}
\alias{timeExpr}
\title{Time the evaluation of an expression}
\usage{
timeExpr(expr, verbose = FALSE)
}
\arguments{
\item{expr}{An expression to be evaluated.}

\item{verbose}{Whether to display the elapsed time as output.}
}
\value{
A numeric value representing elapsed time.
}
\description{
This function computes the time required to evaluate an expression.
}
\examples{
timeExpr(sum(2, 3, 4))

}
