% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timeFun}
\alias{timeFun}
\title{Time a function}
\usage{
timeFun(fun, ...)
}
\arguments{
\item{fun}{A function.}

\item{...}{Additional parameters passed to the function.}
}
\value{
The function output.
}
\description{
This function prints the time required to run a function and returns the
function output.
}
\examples{
res <- timeFun(sum, 2, 3, 4)

}
