% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_and_memory.R
\name{timeMemoryExpr}
\alias{timeMemoryExpr}
\title{Compute the elapsed time, memory and peak memory required to evaluate an
expression}
\usage{
timeMemoryExpr(expr, verbose = FALSE)
}
\arguments{
\item{expr}{An expression to be evaluated.}

\item{verbose}{Whether to display the elapsed time as output.}
}
\value{
A numeric vector containing the elapsed time, the memory usage and
the peak memory usage.
}
\description{
This function computes the elapsed time, memory and peak memory required to
evaluate an expression.
}
\examples{
timeMemoryExpr(sum(2, 3, 4))

}
