% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_object.R
\name{addCategory}
\alias{addCategory}
\title{Add a categorical column to a data frame based on another column}
\usage{
addCategory(df, col, newCol, hashKeys, hashValues)
}
\arguments{
\item{df}{A data frame.}

\item{col}{Column whose values will be used for creating the new column.}

\item{newCol}{Column to be added.}

\item{hashKeys}{A list of hash keys. If vectors are part of the hash keys,
each of their elements will be assigned the corresponding value.}

\item{hashValues}{A vector of hash values.
Must have the same length as \code{hashKeys}.}
}
\value{
A data frame with a new categorical column.
}
\description{
This function adds a categorical column to a data frame based on another
column.
}
\examples{
df <- data.frame(fruit = c('apple', 'banana', 'cherry', 'grape'))
df <- addCategory(df,
                'fruit',
                'color',
                list(c('apple', 'cherry'),
                'banana',
                'grape'),
                c('red', 'yellow', 'purple'))
}
