% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_object.R
\name{addMetadataCategory}
\alias{addMetadataCategory}
\title{Add a categorical column to a Seurat metadata or SingleCellExperiment
coldata}
\usage{
addMetadataCategory(
  scObj,
  col,
  newCol,
  hashKeys,
  hashValues,
  newCol2 = NULL,
  hashValues2 = NULL
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{col}{Column whose values will be used for creating the new column.}

\item{newCol}{Column to be added.}

\item{hashKeys}{A list of hash keys. If vectors are part of the hash keys,
each of their elements will be assigned the corresponding value.}

\item{hashValues}{A vector of hash values.
Must have the same length as \code{hashKeys}.}

\item{newCol2}{A second column to be added based on the same hash keys.
Default is \code{NULL} (no second column will be added).}

\item{hashValues2}{A vector of hash values corresponding to the second
column. Default is \code{NULL} (no second column will be added).}
}
\value{
A Seurat or SingleCellExpression object with one or two new
categorical column(s) in the metadata/coldata.
}
\description{
Add a categorical column to a Seurat metadata or SingleCellExperiment
coldata
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
sceObj <- addMetadataCategory(sceObj, 'Cell_Cycle', 'Type',
list(c('G0', 'G1'), 'G2M', 'S'), c(2, 3, 1))

}
