% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{addNormSilhouette}
\alias{addNormSilhouette}
\title{Adds normalized silhouette column to a single-cell expression object}
\usage{
addNormSilhouette(scObj, normSilDF, normSilCol = "normSilhouette")
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{normSilDF}{Normalized silhouette data frame.}

\item{normSilCol}{The name of the normalized silhouette column to be added.}
}
\value{
The input object (\code{Seurat} or \code{SingleCellExperiment}) with
an added metadata normalized silhouette column.
}
\description{
This function adds a normalized silhouette column to a single-cell
expression object.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
df <- normalizeSilhouette(sceObj, 'Cluster')
sceObj <- addNormSilhouette(sceObj, df)

}
