% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{colsDimPlot}
\alias{colsDimPlot}
\title{Plot Seurat DimPlot with added labeled points for numeric columns}
\usage{
colsDimPlot(seuratObj, cols, ...)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{cols}{Genes whose centers of mass will be plotted.}

\item{...}{Additional parameters passed to \code{pointsDimPlot}.}
}
\value{
A ggplot object.
}
\description{
This function plots a Seurat DimPlot with added labeled points for metadata
numeric columns.
}
\examples{
seuratPath <- system.file('extdata', 'seuratObj.qs', package='hammers')
seuratObj <- qs::qread(seuratPath)
colsDimPlot(seuratObj, c('nCount_originalexp', 'nFeature_originalexp'))

}
