% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{computeSilhouette}
\alias{computeSilhouette}
\title{Compute cluster silhouette for single-cell expression object}
\usage{
computeSilhouette(scObj, idClass, pcaMat = NULL)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{idClass}{Identity class. Must be present among the metadata columns of
the single-cell expression object.}

\item{pcaMat}{PCA matrix.}
}
\value{
The input object (\code{Seurat} or \code{SingleCellExperiment}) with
an added metadata silhouette column.
}
\description{
This function computes the silhouette for each cell in the \code{Seurat}
or \code{SingleCellExperiment} object.
}
