% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{computeSilhouette}
\alias{computeSilhouette}
\title{Compute cluster silhouette for single-cell expression object}
\usage{
computeSilhouette(scObj, idClass, silCol = "silhouette")
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{idClass}{Identity class. Must be present among the metadata columns of
the single-cell expression object.}

\item{silCol}{The name of the silhouette column to be added.}
}
\value{
The input object (\code{Seurat} or \code{SingleCellExperiment}) with
an added metadata silhouette column.
}
\description{
This function computes the silhouette for each cell in the \code{Seurat}
or \code{SingleCellExperiment} object.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scObj <- scuttle::logNormCounts(scObj)
scObj <- scater::runPCA(scObj)
scCol(scObj, 'Cluster') <- withr::with_seed(1,
sample(paste0('Cluster', seq(5)), dim(scObj)[2], replace=TRUE))
scObj <- computeSilhouette(scObj, 'Cluster')

}
