% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{distributionPlot}
\alias{distributionPlot}
\title{Plot the distribution of cells across two columns}
\usage{
distributionPlot(
  scObj,
  plotTitle = "Distribution plot",
  col1 = seurat_clusters,
  col2 = orig.ident,
  xLab = "Column 1",
  yLab = "Count",
  legendLab = "Column 2",
  palette = "Spectral"
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{plotTitle}{Plot title.}

\item{col1}{Column 1}

\item{col2}{Column 2.}

\item{xLab}{x axis label.}

\item{yLab}{y axis label.}

\item{legendLab}{Legend label.}

\item{palette}{Color palette.}
}
\description{
This function plots the distribution of cells across two columns.
}
