% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_of_mass.R
\name{geneCenters}
\alias{geneCenters}
\title{Calculate the centers of mass of the expression of input genes}
\usage{
geneCenters(scObj, genes)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{genes}{Selected genes. If \code{NULL}, all genes will be retained.}
}
\value{
A data frame containing the centers of mass.
}
\description{
This function calculates the centers of mass of the expression of input
genes.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
geneCenters(sceObj, c('Gene_0480', 'Gene_0481', 'Gene_0482'))

}
