% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_of_mass.R
\name{geneCenters}
\alias{geneCenters}
\title{Calculate the coordinates of centers of mass of gene expression}
\usage{
geneCenters(scObj, genes)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{genes}{Selected genes. If \code{NULL}, all genes will be retained}
}
\value{
A data frame containing the coordinates of centers of mass.
}
\description{
This function calculates the coordinates of the center of mass of the
expression of input genes.
}
\examples{
sceObj <- scRNAseq::BaronPancreasData('human')
sceObj <- scuttle::logNormCounts(sceObj)
sceObj <- scater::runUMAP(sceObj)
geneCenters(sceObj, c('AURKA', 'MKI67', 'TOP2A'))


}
