% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{genesDimPlot}
\alias{genesDimPlot}
\title{Plot Seurat DimPlot with added labeled points for genes}
\usage{
genesDimPlot(seuratObj, genes, ...)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{genes}{Genes whose centers of mass will be plotted.}

\item{...}{Additional parameters passed to \code{pointsDimPlot}.}
}
\value{
A ggplot object.
}
\description{
This function plots a Seurat DimPlot with added labeled points for genes.
}
\examples{
seuratPath <- system.file('extdata', 'seuratObj.qs', package='hammers')
seuratObj <- qs::qread(seuratPath)
genesDimPlot(seuratObj, c('Spike-0021', 'Spike-0053', 'Spike-0018'))

}
