% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{genesDimPlot}
\alias{genesDimPlot}
\title{Plot Seurat DimPlot with added labeled points for genes}
\usage{
genesDimPlot(seuratObj, genes, ...)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{genes}{Genes whose centers of mass will be plotted.}

\item{...}{Additional parameters passed to \code{pointsDimPlot}.}
}
\value{
A ggplot object.
}
\description{
This function plots a Seurat DimPlot with added labeled points for genes.
}
\examples{
sceObj <- scRNAseq::BaronPancreasData('human')
sceObj <- scuttle::logNormCounts(sceObj)
seuratObj <- suppressWarnings(Seurat::as.Seurat(sceObj))
seuratObj <- Seurat::FindVariableFeatures(seuratObj)
seuratObj <- Seurat::ScaleData(seuratObj)
seuratObj <- Seurat::RunPCA(seuratObj)
seuratObj <- suppressWarnings(Seurat::RunUMAP(seuratObj, dims=1:15))
genesDimPlot(seuratObj, c('AURKA', 'TOP2A', 'MKI67'))

}
