% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{genesER}
\alias{genesER}
\title{Perform enrichment analysis on a set of genes}
\usage{
genesER(genes, species, funStr = c("enrichGO", "enrichKEGG", "enrichWP"))
}
\arguments{
\item{genes}{A character vector of gene symbols.}

\item{species}{Species. Must be one of 'human', 'mouse' and 'zebrafish'.}
}
\value{
Enrichment result.
}
\description{
This function performs enrichment analysis on a set of genes.
}
