% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{mtCorrectDF}
\alias{mtCorrectDF}
\title{Perform multiple testing correction}
\usage{
mtCorrectDF(df, mtMethod = c("bf", "bh", "by"), ...)
}
\arguments{
\item{df}{A dataframe with a column of p-values.}

\item{mtMethod}{Multiple testing correction method. Options are
Bonferroni ('bf'), Benjamini-Hochberg('bh'), and Benjamini-Yekutieli ('by').}

\item{...}{Additional arguments passed to the multiple testing correction
method.}
}
\description{
This function performs correction for multiple testing in a dataframe column
of p-values and filters the data-frame based on p-values.
}
\examples{
df <- data.frame(elem = c('A', 'B', 'C', 'D', 'E'),
pval = c(0.032, 0.001, 0.0045, 0.051, 0.048))
mtCorrectDF(df)

}
