% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{normalizeSilhouette}
\alias{normalizeSilhouette}
\title{Normalize silhouette by identity class for single-cell expression object}
\usage{
normalizeSilhouette(scObj, idClass)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{idClass}{Identity class. Must be present among the metadata columns of
the single-cell expression object.}
}
\value{
A data frame with normalized silhouettes for each unique element in
the identity class.
}
\description{
This function normalizes the already computed silhouette for each identity
class in the single-cell expression object.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scObj <- scuttle::logNormCounts(scObj)
scObj <- scater::runPCA(scObj)
scCol(scObj, 'Cluster') <- withr::with_seed(1,
sample(paste0('Cluster', seq(5)), dim(scObj)[2], replace=TRUE))
scObj <- computeSilhouette(scObj, 'Cluster')
df <- normalizeSilhouette(scObj, 'Cluster')

}
