% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{normalizeSilhouette}
\alias{normalizeSilhouette}
\title{Normalize silhouette by identity class for single-cell expression object}
\usage{
normalizeSilhouette(scObj, idClass)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{idClass}{Identity class. Must be present among the metadata columns of
the single-cell expression object.}
}
\value{
A data frame with normalized silhouettes for each unique element in the
identity class.
}
\description{
This function normalizes the already computed silhouette for each identity
class in the single-cell expression object.
}
