% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pointsDimPlot}
\alias{pointsDimPlot}
\title{Plot Seurat DimPlot with added labeled points}
\usage{
pointsDimPlot(
  seuratObj,
  plotTitle = NULL,
  pointsObj = NULL,
  alpha = 1,
  pointShape = 4,
  pointSize = 2,
  pointColor = "black",
  labelSize = 2.5,
  maxOverlaps = 30,
  ...
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{plotTitle}{Plot title.}

\item{pointsObj}{A data frame or matrix of points with two columns
representing x and y coordinates.}

\item{alpha}{Opaqueness level.}

\item{pointShape}{Point shape.}

\item{pointSize}{Point size.}

\item{pointColor}{Point color.}

\item{labelSize}{Label size. If \code{NULL}, the points will not receive
labels.}

\item{maxOverlaps}{Maximum overlaps.}

\item{...}{Additional parameters passed to \code{Seurat::DimPlot}.}
}
\value{
A ggplot object.
}
\description{
This function plots a Seurat DimPlot with added labeled points.
}
\examples{
seuratPath <- system.file('extdata', 'seuratObj.qs', package='hammers')
seuratObj <- qs::qread(seuratPath)
pointsObj <- data.frame(x = c(2, 3),
y = c(1, 0),
row.names = c('P1', 'P2'))
pointsDimPlot(seuratObj, pointsObj=pointsObj)

}
