% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pointsDimPlot}
\alias{pointsDimPlot}
\title{Plot Seurat DimPlot with added labeled points}
\usage{
pointsDimPlot(
  seuratObj,
  plotTitle = "Dim plot",
  pointsDF = NULL,
  pointShape = 4,
  pointSize = 2,
  labelSize = 2.5,
  maxOverlaps = 30,
  ...
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{plotTitle}{Plot title.}

\item{pointsDF}{A data frame of points with two columns representing the x
and y coordinates.}

\item{pointShape}{Point shape.}

\item{pointSize}{Point size.}

\item{labelSize}{Label size.}

\item{maxOverlaps}{Maximum overlaps.}

\item{...}{Additional parameters passed to \code{Seurat::DimPlot}.}
}
\value{
A ggplot object.
}
\description{
This function plots a Seurat DimPlot with added labeled points.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scObj <- suppressWarnings(Seurat::as.Seurat(scObj, data=NULL))
scObj <- Seurat::NormalizeData(scObj)
scObj <- Seurat::FindVariableFeatures(scObj)
scObj <- Seurat::ScaleData(scObj)
scObj <- suppressWarnings(Seurat::RunPCA(scObj))
scObj <- suppressWarnings(Seurat::RunUMAP(scObj, dims=1:15))
pointsDF <- data.frame(x = c(2, 3),
y = c(1, 0),
row.names = c('P1', 'P2'))
pointsDimPlot(scObj, pointsDF=pointsDF)

}
