% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pointsDimPlot}
\alias{pointsDimPlot}
\title{Plot Seurat DimPlot with added labeled points}
\usage{
pointsDimPlot(
  seuratObj,
  plotTitle = NULL,
  pointsDF = NULL,
  pointShape = 4,
  pointSize = 2,
  labelSize = 2.5,
  maxOverlaps = 30,
  ...
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{plotTitle}{Plot title.}

\item{pointsDF}{A data frame of points with two columns representing the x
and y coordinates.}

\item{pointShape}{Point shape.}

\item{pointSize}{Point size.}

\item{labelSize}{Label size.}

\item{maxOverlaps}{Maximum overlaps.}

\item{...}{Additional parameters passed to \code{Seurat::DimPlot}.}
}
\value{
A ggplot object.
}
\description{
This function plots a Seurat DimPlot with added labeled points.
}
\examples{
seuratPath <- system.file('extdata', 'seuratObj.qs', package='hammers')
seuratObj <- qs::qread(seuratPath)
pointsDF <- data.frame(x = c(2, 3),
y = c(1, 0),
row.names = c('P1', 'P2'))
pointsDimPlot(seuratObj, pointsDF=pointsDF)

}
