% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepAlluvial}
\alias{prepAlluvial}
\title{Prepare dataframe for alluvial plot}
\usage{
prepAlluvial(
  df,
  pvalCol = "pvalAdj",
  colIndices = c(1, 2),
  weightExp = 1/2,
  offset = 1.00000023069254e-317
)
}
\arguments{
\item{df}{A data frame.}

\item{pvalCol}{Name of p-value column.}

\item{colIndices}{A vector respresenting the indices of the two categorical
columns from the data frame that will be used.}

\item{weightExp}{Exponent used in constructing weight from p-values.}

\item{offset}{Offset used to avoid zeros inside the logarithm function.}
}
\value{
A data frame with weight scores in lieu of p-values.
}
\description{
This function extracts the relevant information from dataframe and adjusts
p-values to be used as weights for the alluvia.
}
\keyword{internal}
