% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pvalRiverPlot}
\alias{pvalRiverPlot}
\title{Plot representation data frame}
\usage{
pvalRiverPlot(
  df,
  pvalCol = "pvalAdj",
  colIndices = c(1, 2),
  weightExp = 1/2,
  pvalOffset = 1.00000023069254e-317,
  ...
)
}
\arguments{
\item{df}{A data frame.}

\item{pvalCol}{Name of p-value column to be used by the alluvial plot.}

\item{colIndices}{A vector respresenting the indices of the two categorical
columns from the data frame that will be used.}

\item{weightExp}{Exponent used in constructing weight from p-values.}

\item{pvalOffset}{Offset used to avoid zeros inside the logarithm function.}

\item{...}{Additional parameters passed to \code{henna::riverPlot}}
}
\value{
A ggplot object
}
\description{
This function plots representation data frame as an alluvial plot.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
df <- repAnalysis(sceObj, 'Cluster', 'Donor')
pvalRiverPlot(df)

}
