% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pvalRiverPlot}
\alias{pvalRiverPlot}
\title{Plot representation data frame}
\usage{
pvalRiverPlot(df, weightExp = 1/2, ...)
}
\arguments{
\item{df}{A data frame.}

\item{weightExp}{Exponent used in constructing weight from p-values.}

\item{...}{Additional parameters passed to \code{henna::riverPlot}}
}
\value{
A ggplot object
}
\description{
This function plots representation data frame as an alluvial plot.
}
\examples{
scObj <- scRNAseq::BaronPancreasData('human')
df <- repAnalysis(scObj, 'donor', 'label')
pvalRiverPlot(df)

}
