% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_genes.R
\name{removeRareGenes}
\alias{removeRareGenes}
\title{Remove rare genes from a Seurat or SingleCellExpression object}
\usage{
removeRareGenes(scObj, nCells = 10, verbose = TRUE)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{nCells}{Minimum number of cells in which a gene must be expressed to
be retained.}

\item{verbose}{Logical; whether the output should be verbose.}
}
\value{
A Seurat or SingleCellExpression object with rare genes removed.
}
\description{
This function removes genes expressed in a low number of cells in a Seurat
or SingleCellExpression object.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scObj <- removeRareGenes(scObj, 30)

}
