% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{repAnalysis}
\alias{repAnalysis}
\title{Find the differential representation of two Seurat columns}
\usage{
repAnalysis(
  scObj,
  col1 = "seurat_clusters",
  col2 = "orig.ident",
  doOverrep = TRUE,
  fdrMethod = c("BY", "BH"),
  pvalThr = 0.05
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{col1}{Column as string.}

\item{col2}{Column as string.}

\item{doOverrep}{Whether to perform overrepresentation analysis. If
\code{FALSE}, underrepresentation analysis will be performed instead.}

\item{fdrMethod}{False discovery rate control method. Options are 'BY'
(Benjamini-Yekutieli) and 'BH' (Benjamini-Hochberg).}

\item{pvalThr}{p-value threshold.}
}
\value{
An overrepresentation or underrepresentation data frame.
}
\description{
This function find the differential representation of two Seurat columns.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scCol(scObj, 'Cluster') <- withr::with_seed(1,
sample(paste0('Cluster', seq(5)), dim(scObj)[2], replace=TRUE))
scCol(scObj, 'Donor') <- rep('Donor1', dim(scObj)[2])
for (i in seq(5)){
scCol(scObj, 'Donor')[sample(which(scCol(scObj, 'Cluster') ==
paste0('Cluster', i)), 15)]<- paste0('Donor', i)
scCol(scObj, 'Donor')[sample(which(scCol(scObj, 'Cluster') ==
paste0('Cluster', i))
, 15)]<- paste0('Donor', i + 1)
}
repAnalysis(scObj, 'Cluster', 'Donor')

}
