% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{repAnalysis}
\alias{repAnalysis}
\title{Find the differential representation of two Seurat columns}
\usage{
repAnalysis(
  scObj,
  col1 = "seurat_clusters",
  col2 = "orig.ident",
  doOverrep = TRUE,
  mtMethod = c("by", "bf", "bh"),
  ...
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{col1}{Column as string.}

\item{col2}{Column as string.}

\item{doOverrep}{Whether to perform overrepresentation analysis. If
\code{FALSE}, underrepresentation analysis will be performed instead.}

\item{mtMethod}{Multiple testing correction method. Choices are
Bonferroni ('bf'), Benjamini-Hochberg('bh'), and Benjamini-Yekutieli ('by').}

\item{...}{Additional arguments passed to the multiple testing correction
method.}
}
\value{
An overrepresentation or underrepresentation data frame.
}
\description{
This function find the differential representation of two Seurat columns.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='hammers')
sceObj <- qs::qread(scePath)
repAnalysis(sceObj, 'Cluster', 'Donor')

}
