% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{repAnalysis}
\alias{repAnalysis}
\title{Find the differential representation of two Seurat columns}
\usage{
repAnalysis(
  scObj,
  col1 = "seurat_clusters",
  col2 = "orig.ident",
  doOverrep = TRUE,
  pvalThr = 0.05
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{col1}{Column as string.}

\item{col2}{Column as string.}

\item{doOverrep}{Whether to perform overrepresentation analysis. If
\code{FALSE}, underrepresentation analysis will be performed instead.}

\item{pvalThr}{p-value threshold.}
}
\value{
An overrepresentation or underrepresentation data frame.
}
\description{
This function find the differential representation of two Seurat columns.
}
