% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R, R/generics.R
\name{scCol.default}
\alias{scCol.default}
\alias{scCol.Seurat}
\alias{scCol.SingleCellExperiment}
\alias{scCol}
\title{Extract a metadata/coldata column from object.}
\usage{
\method{scCol}{default}(scObj, colStr)

\method{scCol}{Seurat}(scObj, colStr)

\method{scCol}{SingleCellExperiment}(scObj, colStr)

scCol(scObj, col)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{col}{Column name.}
}
\value{
A vector.
}
\description{
This function extracts a metadata/coldata column from a Seurat or
SingleCellExperiment object.
}
\examples{
scObj <- scRNAseq::BaronPancreasData('human')
v <- scCol(scObj, 'label')

}
