% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R, R/generics.R
\name{scPCAMat.default}
\alias{scPCAMat.default}
\alias{scPCAMat.Seurat}
\alias{scPCAMat.SingleCellExperiment}
\alias{scPCAMat}
\title{Extracts the PCA matrix from object.}
\usage{
\method{scPCAMat}{default}(scObj)

\method{scPCAMat}{Seurat}(scObj)

\method{scPCAMat}{SingleCellExperiment}(scObj)

scPCAMat(scObj)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}
}
\value{
A PCA matrix.
}
\description{
This function extracts the PCA matrix from a Seurat or
SingleCellExperiment object.
}
