% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffleGenes}
\alias{shuffleGenes}
\title{Replaces genes from vector}
\usage{
shuffleGenes(
  scObj,
  genes,
  lossFrac,
  noiseFrac,
  geneCountThresh = 10,
  seed = 1,
  verbose = TRUE
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{genes}{A character vector.}

\item{lossFrac}{Fraction of genes than be removed. Must be in \code{[0, 1]}.}

\item{noiseFrac}{Amount of noise (random genes) in the final gene vector.
Must be in \code{[0, 1)}}

\item{geneCountThresh}{Minimum number of cells in which newly added genes
must be expressed.}

\item{seed}{Random seed.}

\item{verbose}{Whether the output should be verbose.}
}
\value{
Genes vector after changes.
}
\description{
This function removes and adds genes from vector at random.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
genes <- c('Gene_0826', 'Gene_0610', 'Gene_0380', 'Gene_0602',
'Gene_0613', 'Gene_0201', 'Gene_0295')
shuffleGenes(scObj, genes, 0.3, 0.9)

}
