% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffleGenes}
\alias{shuffleGenes}
\title{Replaces genes from vector}
\usage{
shuffleGenes(scObj, genes, lossFrac, noiseFrac, geneCountThresh = 10, seed = 1)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{genes}{A character vector.}

\item{lossFrac}{Fraction of genes than be removed. Must be in \code{[0, 1]}.}

\item{noiseFrac}{Amount of noise (random genes) in the final gene vector.
Must be in \code{[0, 1)}}

\item{geneCountThresh}{Minimum number of cells in which newly added genes
must be expressed.}

\item{seed}{Random seed.}
}
\value{
Genes vector after changes.
}
\description{
This function removes and adds genes from vector at random.
}
