% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{termGenes}
\alias{termGenes}
\title{Extract genes enriched for terms}
\usage{
termGenes(er, terms, negTerms = NULL)
}
\arguments{
\item{er}{Enrichment result of the \code{enrichResult} class.}

\item{terms}{Terms for which enriched genes should be extracted.}

\item{negTerms}{Terms for which enriched genes should be subtracted from the
genes enriched for \code{terms}.}
}
\value{
Genes enriched for terms.
}
\description{
This function extracts genes enriched for terms from an \code{enrichResult}
object.
}
\examples{
m <- genesER(c('AURKA', 'TOP2A', 'CENPF', 'PTTG2', 'MKI67', 'BIRC5', 'RRM2'),
'human')
termGenes(m, 'chromosome segregation', 'meiosis I')

}
