% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{termGenes}
\alias{termGenes}
\title{Extract genes enriched for term}
\usage{
termGenes(er, terms, negTerms = NULL)
}
\arguments{
\item{er}{Enrichment result.}

\item{terms}{Terms for which enriched genes should be extracted.}

\item{negTerms}{Terms for which enriched genes should be subtracted from the
genes enriched for \code{terms}.}
}
\value{
Genes enriched for terms.
}
\description{
This function extracts genes enriched for term from an \code{enrichResult}
object.
}
